backend/requirements.txt
requests>=2.31.0
pandas>=2.1.0
sqlalchemy>=2.0.0
tqdm>=4.66.0
urllib3>=2.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0 